feat: change score field from int to float 0.0-1.0
Change the score field throughout the codebase from binary integer
(0 or 1) to a float in the 0.0-1.0 range, enabling partial credit
and confidence-weighted scoring.

Changes:
- GradeResult.score: int -> float, validates 0.0-1.0 range
- TestResult.score: int = -1 -> float = -1.0 (sentinel preserved)
- MultiGradeResult: scores List[int] -> List[float], majority_score float
- DB schema: score INTEGER -> score REAL (SQLite + SQLAlchemy)
- PostgreSQL migration: ALTER COLUMN score TYPE REAL
- db_listener/grader: parse scores as float() instead of int()
- Robot resource: Should Be Equal As Numbers instead of Integers
- import_test_results.py: parse scores as float
- Version bump: 1.3.2 -> 1.3.3

+++ src/rfc/grader.py
@@ -92,6 +92,6 @@ Format:
             raise ValueError(f"Grader JSON missing required fields: {parsed}")
 
         return GradeResult(
-            score=int(parsed["score"]),
+            score=float(parsed["score"]),
             reason=str(parsed["reason"]),
         )
